minimal pagination for upgrade UI
